Home_Task10 1.	Implement the following design patterns in your solution from previous module*: a.	Singleton b.	Factory Method c.	Decorator 2.	Revise your code to match S.O.L.I.D. principles.
