+++ src/test/java/com/epam/tests/UI/AbstractTest.java
@@ -1,6 +1,6 @@
 package com.epam.tests.UI;
 
-import driver.DriverManager;
+import driver.DriverSingleton;
 import org.openqa.selenium.WebDriver;
 import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeMethod;
@@ -16,11 +16,11 @@ public class AbstractTest {
 
     @BeforeMethod
     public void startDriver() {
-        driver = DriverManager.getDriver();
+        driver = DriverSingleton.getDriver();
     }
 
     @AfterMethod
     public void closeBrowser() {
-        DriverManager.closeDriver();
+        DriverSingleton.closeDriver();
     }
 }
